fix: button focus outline
Most of our components do not have an outline border set, so when you tab to them
they have a default blue outline. This changes Button, CloseButton, and Tab to
use their hover colors (purple) instead. Button/close button also require
outline-transparent to avoid a white outline flash on entry.

This fixes the specific issue (buttons), but there are >10 other controls to fix
as well. It's not always clear what the correct behaviour should be, so I've
opened #11482 to track the overall design.

Fixes #6379.

Signed-off-by: Tim deBoer <git@tdeboer.ca>