chore: feedback fixes
Fixes based on PR feedback:
- Property defined in /main using standard pattern.
- Tests updated to use Object.defineProperty().
- Property is now reactive to anything external changing the configuration.

Also updated to Svelte 5, and fixed a bug I noticed during testing: transitions were
always applied when creating the component, so e.g. if it was closed and then you
switch to another view and back you see it closing again. Adding an initialization
property fixes this.

Signed-off-by: Tim deBoer <git@tdeboer.ca>