chore: common kubernetes ui object
First step of #11035 - a common Kubernetes UI object. This just adds some minimal
typing which allows the name column to be shared. Namespace is optional since it
doesn't exist on nodes.

The name column is used for pods as an example before I do the rest. A 'kind'
property was added to map urls.

The rest of #11035 (rest of name columns, column folders/naming, props.ts) would
come in future PRs.

Signed-off-by: Tim deBoer <git@tdeboer.ca>