chore: remove kubernetes pods from pods page
First part of #8819. Removes the data for Kubernetes pods from the Pods page, and
adds a simple floating message to the bottom right for anyone trying to figure out
where their Kube pods are. This message would eventually be replaced with the
recommendations from design in #10812.

This is the core breaking change for the user, removes the watcher code, removes
the internal API, and cleans up the pod store to be consistent with other
container engine stores. Plenty for one PR, but there is still more to do in
future PRs:
- removing actions / if statements that don't apply or will never receive Kube pod
  data now
- likely remove Environment column
- look for more internal cleanup

First part of #8819.

Signed-off-by: Tim deBoer <git@tdeboer.ca>