chore: remove kubernetes pods from pods page
First part of #8819. Removes the data for Kubernetes pods from the Pods page, and
adds a simple message to the top for anyone trying to figure out where their Kube
pods are.

This is the core initial/minimal breaking change for the user, without any of the
internal or UI cleanup.

First part of #8819.

Signed-off-by: Tim deBoer <git@tdeboer.ca>