feat: kubernetes resource navigation
Splits ingress/routes and configmap/secret into separate functions to be consistent
with every other resource, and adds navigation for any supported resource via its
kind.

Since all existing paths use the function, KUBERNETES_RESOURCES / KUBERNETES_RESOURCE
is being implicitly tested by all of the existing tests for correct urls.

There is an argument to be made that once we've added KUBERNETES_RESOURCES /
KUBERNETES_RESOURCE we don't need any of the previous Kubernetes paths because it's easy
enough to add { kind: 'DEPLOYMENT' } to any call, but I've left them in for simplicity.
If anyone feels strongly I can remove them.

gotoKubernetesResource() and resourceKindToURL() encode the 4 (really 6) urls that don't
follow the normal Kubernetes convention. There's nothing inherently wrong with that and
we could reduce it, but it is now easier to see and should be a separate issue.

Signed-off-by: Tim deBoer <git@tdeboer.ca>