chore: remove kubernetes from internal navigation api
Now that we have the Kubernetes navigation route (#11166) to any Kubernetes
resource, we do not need the Kubernetes navigation in the internal navigation
API. Since these could cause confusion and would be removed when/if Kubernetes
is moved to a built-in extension, they should be removed now. (yes, even the
bits I merged last week)

I can't find any use of these internally (?), and they are not exposed to
extensions.

Part of #11219.

Signed-off-by: Tim deBoer <git@tdeboer.ca>