chore: common function for kubernetes icons
Every Kubernetes resource kind has an associated icon, so when we know the kind
why not look it up vs having to specify the icon everywhere? This creates a utility
function in kube-utils to return the icon based on kind, and uses that in the
dashboard resource card to remove the property.

Isn't this just moving this function? Sure. But it means we can now look up the
icon *anywhere* we know the kind, e.g. anywhere the UI objects exist. The next
PR can create a shared status column and remove the individual status columns
from all lists. We might be able to do something similar with the empty pages.

Part of #11035.

Signed-off-by: Tim deBoer <git@tdeboer.ca>