chore: kubernetes icon component
Every Kubernetes resource kind has an associated icon, so when we know the kind
why not use a component vs specifying the icon everywhere? This creates a new
KubernetesIcon that is used by the dashboard resource card but could also be used
by empty pages and status columns in the future.

There is no easy way to test that a rendered svg matches, so I added a test id
to every Kubernetes icon with the resource kind(s) that it represents.

Part of #11035.

Signed-off-by: Tim deBoer <git@tdeboer.ca>