fix: add back button cursor pointer
Reverts the Tailwind v4 change to remove the default button cursor pointer.

As discussed on the standup, we'll undo this change for the current release
to avoid regression (especially on components that have no hover effect),
then possibly undo this later and replace with targeted fixes for elements
where we have no hover or focus indicator.

Fixes #11436.

Signed-off-by: Tim deBoer <git@tdeboer.ca>