chore: clean up PodInfoUI kind
The kind property in PodInfoUI can only be 'podman', so remove it.

Likewise, PodGroupInfoTypeUI is only used by ProviderInfo and can only be
'podman'.

ProviderInfo could probably remove all Kubernetes support at this point,
but since the 'kubernetes' string is already used in the file and the file
is used in lots of other places I'm just doing the basic removal of the
type here.

Part of #8819.

Signed-off-by: Tim deBoer <git@tdeboer.ca>