chore: api to change kubernetes namespace
Adds new function to allow the user to change the Kubernetes namespace on the
current context and updates state, but without modifying the kubeconfig. Logic
is similar to what happens in the refresh() method, and exposed to renderer.

Existing function on the TestKubernetesClient renamed due to conflict and test
confirms that the namespace is updated and event fired without modifying
kubeconfig.

This will support a UI allowing the user to quickly switch namespaces from
Kubernetes resource pages.

First part of #11274.

Signed-off-by: Tim deBoer <git@tdeboer.ca>