chore: clone kubeconfig instead of re-reading from disk
syncResources() (apply yaml) can apply to any context or namespace, so it needs
a copy of the kubeconfig to avoid changing the user's current context. However,
instead of re-reading the file from disk it can just clone the config, which
should be marginally faster and safer. This code is copied directly from the
same logic in the refresh() function.

Found during review for #11753.

Signed-off-by: Tim deBoer <git@tdeboer.ca>