chore: update ui package icons to svelte 5
Updates StarIcon and ContainerIcon to Svelte 5 syntax.

The event store type in renderer has to be updated to avoid type errors. Note
that ContainerIcon size and solid had to be made optional ('?') as well (the
other option seemed to be using 'Component<any>'), could be cleaned up later
but probably not worth tracking.

Removing import/no-duplicates as it doesn't seem to be required anymore.

Fixes #11577.

Signed-off-by: Tim deBoer <git@tdeboer.ca>