fix: name column alignment
Moving to Electron 35 changed the default alignment of text to center, which only
causes UI change when the parent component is wider than the text. So far I've
only found impact to two name columns. I can't find the 'breaking' change listed
anywhere, but it is easy enough to fix with text-left for these two columns.

Fixes #11831.

Signed-off-by: Tim deBoer <git@tdeboer.ca>