chore: migrate tooltip to svelte 5
Migrates the tooltip component to Svelte 5.

The original Tooltip had both a property and a slot with the same name, 'tip'.
This isn't possible in Svelte 5 so I changed the slot to a snippet 'tipSnippet'.

There was a typo in PreferenceCliTool, but it was clearly expecting the component
to support the aria-label property.

Fixes #11587.

Signed-off-by: Tim deBoer <git@tdeboer.ca>