chore: migrate status icon to svelte 5
Migrate the status icon component to Svelte 5.

svelte:component has been removed. In order to support it being string | Component
(and needing capitalization in the second case), I've used the technique provided
by the Svelte team here:
https://github.com/sveltejs/svelte/issues/12668#issuecomment-2563451764

Typing of icon could be improved but deferring that as it is a bigger change and
not related to Svelte migration.

Fixes #11584.

Signed-off-by: Tim deBoer <git@tdeboer.ca>