chore: image and manifest details navigation
Two prior PRs fixed navigation to image details and added navigation to manifests.
This just switches the links from the Images list Name column to use the navigation
instead of directly using the router - minor cleanup and confirms the nav API works.

Fixed tests to match, including repotag that was inconsistent with the name
and tag, and checking where manifests links to.

Follow up from #12139, #12143.

Signed-off-by: Tim deBoer <git@tdeboer.ca>