chore: fix argos screenshots github button
### What does this PR do?

Hides the "badge" section of the GitHub stars so that we do not have
false errors within Argos if the amount of stars in the repo changed.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/podman-desktop/podman-desktop/issues/12446

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

Tests should pass now

Signed-off-by: Charlie Drage <charlie@charliedrage.com>