feat(extension): add search by description in extension list
I tried searching for the bootc extension by typing `bootc` in the
search box, but I couldn't find it. I realized that the search only
looks at the display name of the extension, which doesn't contain the
word bootc.

I changed the search logic to include the description of the extension
in the search. Now, when I type `bootc`, it matches the extension with
the display name "Bootable Container" because the description mentions
bootc.

Searching only by name seems a bit strict, especially considering
the currently small amount of extensions.

Ideally in the future, extensions will be able to provide a list of
keywords associated with them so we can filter against those as well,
instead of something arbitrary like the description.

Signed-off-by: Omer Tuchfeld <omer@tuchfeld.dev>