chore: correctly type guard e.target for micromark listener
### What does this PR do?

This PR refactors the `micromark-listener-handler` file that handles the
"opening a command" implementation for onboarding.

When this was refactored as the result of
https://github.com/podman-desktop/podman-desktop/pull/11038
`eventTarget` was using type-narrowing manually, which does not allow
accessibility to properties like dataset, disabled, etc which is needed
for the "command" functionality to work.

This PR fixes that by safely checking / type guarding e.target.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/podman-desktop/podman-desktop/issues/12211

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

1. Start onboarding
2. Go to Compose or Kubectl
3. Click the "Want to download a different version?" button and see that
   it works correctly now with no errors.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>