chore: when retrieving a configuration value, check locked and managed-by values
### What does this PR do?

When retrieving a configuration value from within Podman Desktop API, if
that value is "locked" by the "locked.json' file, then it will:

1. Check if the value exists in the locked.json file
2. Check the "default-settings" json file
3. Return the default-settings json file

If there is no value in the default-settings file (went missing or
something, despite the locked.json existing), then it will return the
users default value, which is intentional, as we do not want to return
null / undefined values just because the value doesn't exist.

There is little code added, but A LOT of tests added in order to cover
every possible edge case.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A, internal usage when using locked.json and default-settings.json
files.

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/podman-desktop/podman-desktop/issues/13897

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature
Signed-off-by: Charlie Drage <charlie@charliedrage.com>