chore: add disabled prop to preference item formats
### What does this PR do?

When implementing adding labels into the Preferences page I was unable
to pass in any "disabled" parameters to the components rendered by
preferences.

This PR adds the ability to pass in "disabled" into the UI, by adding
the `disabled` prop to PreferencesRenderingItemFormat.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes #14624

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

OR.

Pass in "disabled=true" and try and click on a component (ex.
`packages/renderer/src/lib/preferences/item-formats/StringItem.svelte`)

Signed-off-by: Charlie Drage <charlie@charliedrage.com>