chore: add task issue template
We have kind/* labels for task, release, and technical debt, but our templates don't
cover any of these - you have to create a feature or a bug and then change the type.
This takes longer and leaves a bunch of odd fields in the issue, e.g. 'Is your
feature request related to a problem?'

We're switching to use Github issue types which also includes these three types.

It should be easier to create one of these issue types, but we also don't want to
clutter up the template list with lots of 'internal' types that will make it more
confusing to external users. The proposal is just to add Task, which is generic
enough and easier to switch to Technical debt or Release.

Part of #14777.

Signed-off-by: Tim deBoer <git@tdeboer.ca>