feat: apply user default registries to Podman Desktop and Podman
## What does this PR do?
Uses the `registries.defaults` configuration value from `settings.json` to get the user defaults registries and apply them to Podman Desktop and Podman machines.

`registries.defaults` are used both in the UI, to "suggest" the default registries, and the Podman extension to directly add them to the `registries.conf.d` file.

In the UI, suggested registries that are block have their `configure` button disabled and any duplicate suggested registries (same prefix) have a log in the console about the duplication and only the existing one is shown in the UI with the blocked value of the user default registry.

In the Podman extension, the user default registries are added to the `registries.conf.d` file with conflict resolution focusing on registries with the same prefix => different blocked/insecure/location value will show a warning in the console and will not add the newest registry. If all the value are the same, the newer registry's mirrors will be added to the ones of the existing registry.

**NOTE:** Feel free to add any additional ideas/thoughts/feedback about the conflict resolution in this PR as there might be a lot of edge cases for duplicate registries (what do we consider as duplicate registry? prefix? location? both?) and ways to deals with them (logs, warnings, errors, overriding, etc.)

### Screenshot / video of UI

N/A

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?
Closes https://github.com/podman-desktop/podman-desktop/issues/13902

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

1. Add to `settings.json` `"registries.defaults": []`
2. Inside the array, add different registry entry following:
```
{
  "registry": {
    "prefix": string,
    "location": string,
    "blocked": boolean | undefined,
    "insecure": boolean | undefined
  }
}
```
or
```
{
  "registry.mirror": {
    "location": string,
    "insecure": boolean | undefined,
  }
}
```
Each mirror for a registry has to come right after the registry entry.

**EDIT BY CHARLIE:**
For a complete example of `settings.json`:

```json
{
  "registries.defaults": [
    {
      "registry": {
        "prefix": "quay.io",
        "location": "quay.io"
      }
    },
    {
      "registry.mirror": {
        "location": "mirror.example.com"
      }
    }
  ]
}
```

3. Upon Podman Desktop start, check that the correct warning/logs are outputted in case of conflicts
4. Check that in the `$HOME/.config/containers/registries.conf` file, the registries have been added correctly and that conflicts have been resolved as expected:

Example:
```
~ $ cat .config/containers/registries.conf
[[registry]]
prefix = "quay.io"
location = "quay.io"

[[registry.mirror]]
location = "mirror.example.com"⏎
```

- [x] Tests are covering the bug fix or the new feature

Signed-off-by: Sonia Sandler <ssandler@redhat.com>
Co-authored-by: Charlie Drage <charlie@charliedrage.com>
Signed-off-by: Charlie Drage <charlie@charliedrage.com>