chore: add more tests, make sure we saveDefault if changes.
Adds more tests to cover the use-case of making sure that if a value is
already default in the schema, it does NOT copy over to setings.json (no
point to have it there... as it's already set by default).

We also make sure that we call saveDefault if there are any
configuration changes. In the previous commit, it would only save if
the user explicitly changes it (saveDefault is then called, and then
checks the schema and saves as necessary).

Signed-off-by: Charlie Drage <charlie@charliedrage.com>