fix: simplified update image to latest build backend
while testing the ui integration for this feature I noticed
that the error for localhost was calling image pull wrongfully
we want to return early if localhost is part of the image handle
and have a clear message for this. We also didnt need to check
the repo digest for population since its only a docker thing
and makes the backend more complex. Really the goal is that
we want to pull an image that can be pulled from remote registry
and delete the old one.

Signed-off-by: Brian <bmahabir@bu.edu>