chore: align telemetry text
I noticed the telemetry text on the Welcome page and preferences have diverged slightly.
They should be the same, especially if we may allow it to be customized in the future
(should be one string).

I could have copy/pasted in either direction, but instead I used the best parts of both,
and confirmed with others that this is the most natural and readable version.

Related to #15379.

Signed-off-by: Tim deBoer <git@tdeboer.ca>