chore(docs): add codesign instructions for macOS
### What does this PR do?

When testing a compiled binary on macOS, it is a special case as there
is validation against the Electron binary before launching.

If you are using `pnpm compile:current`, you should be able to run the
app after launch.

Unfortunatley this requires:

```sh
codesign --force --deep --sign - "dist/mac-arm64/Podman Desktop.app"
```

To be ran.

This adds those instructions to the documentation.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

1. Be on macOS
2. `pnpm compile:current`
3. Try to launch the binary (`dist/mac-arm64/Podman Desktop.app`)
4. Get error. Close app.
5. `codesign --force --deep --sign - "dist/mac-arm64/Podman
   Desktop.app"`
6. No more error when launching.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>